Cmdlet Improvements
Added the option for the user to pass a bookmark path and/or destination folder path to the command to change the file that is being backed up or to change where it is backed up to. 
Changed most of the output text to verbose so running the command only tells you success or failure now.
Simplified failure output text.
Changed format of the time in the output file to be more user-friendly.
Cleaned up the New-Item command that creates the backup folder.
Formatted function names to standard case.
Replaced conditional use of Get-Item with Test-Path which does not produce an error message when it evaluates to false.